% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bestBandwith.R
\name{bestBandwidth}
\alias{bestBandwidth}
\title{Find the best bandwidth.}
\usage{
bestBandwidth(fittedBandwidth, fittedFaults, faults)
}
\arguments{
\item{fittedBandwidth:}{a vector containing the fitted bandwidths.}

\item{fitttedFaults:}{a vector containing the numbers of faults bandwidths was computed for.}

\item{faults:}{the amount of faults on the wafer.}
}
\value{
the best bandwidth.
}
\description{
Return the best bandwidth from the vector of the fitted bandwidth and
the number of faults on the wafer. Another vector should be passed to the function:
it is assumed to contain the amounts of faults every fitted bandwidth was
computed for
}

