% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bindCircularMap.R
\name{bindCircularMap}
\alias{bindCircularMap}
\title{Create a circular fault grid}
\usage{
bindCircularMap(rectangularMap, ray, outValue = NA)
}
\arguments{
\item{rectangularMap:}{a sqare matrix}

\item{ray:}{the ray of the circular map}

\item{outValue:}{the value assignad to the element of of the circle. Default is NA}
}
\value{
the modified map
}
\description{
Creates a circular grid from a square one by inserting the value outValue in
every pixel which is not whitin the circle of given ray.
}

