% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bindDefectNumber.R
\name{bindDefectNumber}
\alias{bindDefectNumber}
\title{Create a map with a specified number of faults}
\usage{
bindDefectNumber(matrix, faultValue, notFaultValue, faultNumber)
}
\arguments{
\item{matrix:}{the matrix filled with faulty and good chips}

\item{faultValue:}{the value which identifies a faulty chip}

\item{notFaultValue:}{the value which identifies a good (not faulty) chip}

\item{faultNumber:}{the number of fault in the matrix which is returned}
}
\value{
a matrix where exactly faultNumber elements are set to faultValue.
}
\description{
Accept a matrix which is assumed to represent the map of faulty and
good chips on a wafer. The fuction returns the same matrix where only
the specified number of element is set to the value which represent a
fauly chip. In other words, this function samples a subset of faulty
chips and saves the other. This is the reason why it is impossible to
specify a number of faulty chip greater then the one of the matrix
passed.
}

