% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chiTest.R
\name{chiTest}
\alias{chiTest}
\title{Find weighed average square error.}
\usage{
chiTest(trueMatrix, extimatedMatrix, na.rm = TRUE)
}
\arguments{
\item{trueMatrix:}{the real matrix.}

\item{extiamtedMatrix:}{the extimated matrix.}

\item{na.rm:}{if set to TRUE (as default), missing values are ignored}
}
\value{
the weighed average square error.
}
\description{
Computes average square error between an extimated matrix and the real
one according to the chi test formula:
error = 1/N(trueMatrix - extimatedMatrix)^2/trueMatrix.
}

