% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/faultNumber.R
\name{faultNumber}
\alias{faultNumber}
\title{Find total amount of faults}
\usage{
faultNumber(faultMap, faultValue)
}
\arguments{
\item{faultMap:}{a matrix containing the simulated chips}

\item{faultValue:}{the value assigned to a faulty chip. It cannot be NA nor NAN}
}
\value{
the total amount of faulty chips.
}
\description{
Computes the total number of faults in a wafer.
}

