% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fillRectangularMap.R
\name{fillRectangularMap}
\alias{fillRectangularMap}
\title{Fills a grid whith two values according to the probabilityFuncion
The higher is the value probabilityFunction takes in a cetrain point,the
higher is the probability that pixel is set to faultValue. The probabilityFuncion
is rescaled so that its maximum value is equal to maxFaultProbability.}
\usage{
fillRectangularMap(probabilityFunction, maxFaultProbability,
  faultValue = TRUE, notFaultValue = FALSE)
}
\arguments{
\item{probabilityFunction:}{a sqare matrix the probability distribution of faults}

\item{maxFaultProbability:}{a real positive number which represents the maximum
probability of a fault}

\item{faultVaue:}{the value assigned to a element which is faulty. Default is TRUE}

\item{notFaultVaue:}{the value assigned to a element which is faulty. Default is FALSE.}
}
\value{
a matrix of the same dimensions of probabilityFunction.
}
\description{
Fills a grid whith two values according to the probabilityFuncion
The higher is the value probabilityFunction takes in a cetrain point,the
higher is the probability that pixel is set to faultValue. The probabilityFuncion
is rescaled so that its maximum value is equal to maxFaultProbability.
}

