% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNormalizationCoefficient.R
\name{getNormalizationCoefficient}
\alias{getNormalizationCoefficient}
\title{Return the normalization coefficient}
\usage{
getNormalizationCoefficient(f, lowerLimit, upperLimit, ...)
}
\arguments{
\item{f:}{the function to evaluate}

\item{lowerLimit:}{a vector containing the lower limits of the region}

\item{upperLimits:}{a vector containing the lower limits of the region}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
the normalization coefficient
}
\description{
Find the normalization coefficient for a given function in a given rectangular
region. In other words, this function returns the coefficient which is able to
make a function a probability function on a given region by normalizing its volume
to 1.
}
\seealso{
adaptIntegrate
}

