% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanSquareError.R
\name{meanSquareError}
\alias{meanSquareError}
\title{Find average square error.}
\usage{
meanSquareError(matrix1, matrix2, na.rm = TRUE)
}
\arguments{
\item{matrix1:}{the first matrix.}

\item{matrix1:}{the second matrix.}

\item{na.rm:}{if set to TRUE (as default), missing values are ignored}
}
\value{
the average square error.
}
\description{
Computes average square error between two matrix.
}

