% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiGaussianDensity.R
\name{multiGaussianDensity}
\alias{multiGaussianDensity}
\title{Return a the normalized sum of bivariate gaussian density}
\usage{
multiGaussianDensity(ray, parameterList)
}
\arguments{
\item{ray:}{half the number of column of the square matrix returned}

\item{parameterList:}{a list whose elements shuold be list, too. They are
assumed to contain the mean and the variace matrix of a single bivariate
normal function}
}
\value{
a list with 2 elemets: the values of the probability function "pdf" and the
grid "grid" where it has been computed.
}
\description{
Computes the sum of an arbitrary number of bivariate gaussian probability
density and return a list containing the grid used to compute the
probability and a matrix with the values of the probability function.
The values in the matrix are normalized to let the volume of the whole
system be equal to 1.
}

