% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMatrix.R
\name{plotMatrix}
\alias{plotMatrix}
\title{Plot a bidimensional matrix}
\usage{
plotMatrix(title = "", matrix, colorMap, ...)
}
\arguments{
\item{title:}{the tithe of the plot}

\item{matrix:}{the matrix to be plotted}

\item{colorMap:}{the colors to be used in the plot. Must be a valid argument for "col".}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Plot a bidimensional matrix using the given colorMap. The title and other arguments
may also be passed to te underlying plot function
}
\seealso{
plot3D
}

