% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\title{Plot a tidimensional surface}
\usage{
plotSurface(title = "", x, y, z, ...)
}
\arguments{
\item{title:}{the tithe of the plot}

\item{x:}{values on the x axis.}

\item{y:}{values on the y axis.}

\item{z:}{values on the z axis - ie the value of the function at the coordinates (x,y).}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Plot a tridimensional surface using the given parameter. The title and other arguments
may also be passed to te underlying plot function.
}
\seealso{
plot3D
}

