% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussianDensity.R
\name{gaussianDensity}
\alias{gaussianDensity}
\title{Gaussian Density}
\usage{
gaussianDensity(x1, x2, mu, sigma)
}
\arguments{
\item{x1:}{sampling points on the x axes}

\item{x2:}{sampling points on the x axes}

\item{mu:}{a vector containing the mean of the gaussian density}

\item{signma:}{a matrix containing the variance of the gaussian density}
}
\value{
a matrix containing the values of the density.
}
\description{
Computes a bivariate gaussian probability density and return the result
in the form of a matrix.
}

